requirements.txt
django>=5.1,<7.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-filter>=24.0
django-cors-headers>=4.3
drf-spectacular>=0.27
python-decouple>=3.8
